requirements.txt
absl-py==0.9.0
alabaster==0.7.12
appnope==0.1.0
astroid @ file:///opt/concourse/worker/volumes/live/b22b518b-f584-4586-5ee9-55bfa4fca96e/volume/astroid_1592495912194/work
attrs==19.3.0
av==8.0.2
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall==0.2.0
black==22.12.0
brotlipy==0.7.0
cachetools==4.1.0
certifi==2021.5.30
cffi @ file:///opt/concourse/worker/volumes/live/ca3c0fea-9d0f-4489-6889-89b42078ed37/volume/cffi_1613246931142/work
chardet @ file:///opt/concourse/worker/volumes/live/9efbf151-b45b-463d-6340-a5c399bf00b7/volume/chardet_1607706825988/work
click==8.1.3
configparser==5.0.0
cryptography @ file:///opt/concourse/worker/volumes/live/cdf8ff17-b0bf-4081-524c-b5a0afe929ba/volume/cryptography_1616769280208/work
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils @ file:///opt/concourse/worker/volumes/live/4d08a32e-3c8a-4d2d-4b17-f3c8b943d07b/volume/docutils_1620827969351/work
fett==0.3.2
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
gql==0.2.0
graphql-core==1.1
grpcio==1.30.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imagesize @ file:///home/ktietz/src/ci/imagesize_1611921604382/work
importlib-metadata==1.6.1
ipykernel @ file:///opt/concourse/worker/volumes/live/73e8766c-12c3-4f76-62a6-3dea9a7da5b7/volume/ipykernel_1596206701501/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///opt/concourse/worker/volumes/live/7ea70a71-5624-4799-50b3-3b90e07922f3/volume/ipython_1593447385860/work
ipython-genutils==0.2.0
isort==4.3.21
jedi @ file:///opt/concourse/worker/volumes/live/114a62c0-dd60-4a40-6382-85abb31b5f6e/volume/jedi_1592841953086/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1624781299557/work
joblib==0.15.1
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe @ file:///opt/concourse/worker/volumes/live/2165c944-5d03-464f-57c1-1df4e51b3ac3/volume/markupsafe_1621528146825/work
matplotlib==3.2.2
mccabe==0.6.1
more-itertools==8.4.0
mypy==0.991
mypy-extensions==0.4.3
networkx==2.5.1
numpy==1.18.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.3.0.36
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.1.0
parso==0.7.0
pathspec==0.10.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
platformdirs==2.6.0
plotly==4.9.0
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.6.1
pylint @ file:///opt/concourse/worker/volumes/live/42ede439-2571-4cb2-513c-394625d2381b/volume/pylint_1592496039330/work
pymongo==3.11.4
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks @ file:///opt/concourse/worker/volumes/live/ef943889-94fc-4539-798d-461c60b77804/volume/pysocks_1605305801690/work
pytest==5.4.3
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
pytorch-lightning==0.8.4
pytube3==9.6.4
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.3.1
pyzmq==19.0.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.17.0
rsa==4.6
rstcheck==3.3.1
scikit-learn==0.23.1
scipy==1.5.0
sentry-sdk==0.15.1
shortuuid==1.0.1
six==1.15.0
sk-video==1.1.10
smmap==3.0.4
snooty-lextudio==1.9.2.dev0
snowballstemmer @ file:///tmp/build/80754af9/snowballstemmer_1611258885636/work
Sphinx @ file:///tmp/build/80754af9/sphinx_1623884544367/work
sphinxcontrib-applehelp @ file:///home/ktietz/src/ci/sphinxcontrib-applehelp_1611920841464/work
sphinxcontrib-devhelp @ file:///home/ktietz/src/ci/sphinxcontrib-devhelp_1611920923094/work
sphinxcontrib-htmlhelp @ file:///tmp/build/80754af9/sphinxcontrib-htmlhelp_1623945626792/work
sphinxcontrib-jsmath @ file:///home/ktietz/src/ci/sphinxcontrib-jsmath_1611920942228/work
sphinxcontrib-qthelp @ file:///home/ktietz/src/ci/sphinxcontrib-qthelp_1611921055322/work
sphinxcontrib-serializinghtml @ file:///tmp/build/80754af9/sphinxcontrib-serializinghtml_1624451540180/work
subprocess32==3.5.4
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
threadpoolctl==2.1.0
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
tomli==2.0.1
torch==1.6.0
torch-tb-profiler==0.2.0
torchvision==0.6.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
typed-ast==1.5.4
typing-extensions==4.4.0
ujson==1.35
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
wandb==0.9.1
watchdog==1.0.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug==1.0.1
wrapt==1.11.2
youtube-dl==2020.6.16.1
zipp==3.1.0
